+++ src/main/java/com/pablosipac/kinalapp/service/ClienteService.java
@@ -105,7 +105,7 @@ public class ClienteService implements IClienteService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<Cliente> listarPorEstado(int estado) {
+    public List<Cliente> listarPorEstado(Long estado) {
         return clienteRepository.findByEstado(1);
     }
